2024-03-03 예약 & 예약취소 기능 JPA 로 수정 (테스트 필요) 23:18

+++ src/main/java/com/green/hoteldog/reservation/model/HotelReservationDelDto.java
@@ -13,9 +13,9 @@ import java.util.List;
 @Schema(title = "호텔 예약 삭제 DTO")
 public class HotelReservationDelDto {
     @Schema(title = "예약 pk")
-    private int resPk;
+    private long resPk;
     @JsonIgnore
-    private int userPk;
+    private long userPk;
     @JsonIgnore
     private List<Integer> resDogPkList = new ArrayList<>();
     @JsonIgnore
